Don't rely on health resurrect for the colony logic.
