+++ OpenRA.Mods.OpenSA/Traits/Colony/Colony.cs
@@ -1,131 +1,59 @@
 using System.Collections.Generic;
 using System.Linq;
-using OpenRA.GameRules;
-using OpenRA.Mods.Common.Effects;
+using OpenRA.Mods.Common;
 using OpenRA.Mods.Common.Traits;
+using OpenRA.Primitives;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.OpenSA.Traits
 {
-	public enum ExplosionType { Footprint, CenterPosition }
-
-	public class ColonyInfo : TraitInfo, Requires<HealthInfo>, IRulesetLoaded
+	public class ColonyInfo : TraitInfo, Requires<TurretedInfo>
 	{
-		public readonly int BitFireDelay = 50;
-		public readonly int NumberOfBits = 8;
-		public readonly int MinRange = 5;
-		public readonly int MaxRange = 10;
-		public readonly int ResurrectHealth = 10;
-
-		[WeaponReference]
-		public readonly string Weapon = "colony_bit";
-
-		[WeaponReference]
-		public readonly string Explode = "dieBuilding";
-
-		public readonly string CaptureSound = "sounds|POWERUP.SDF";
 		public readonly string LostSound = "sounds|POWERDOWN.SDF";
-		public readonly string ColonyExplosionSound = "sounds|COLONYEXPLODE.SDF";
+
+		[ActorReference(typeof(DefeatedColonyInfo))]
+		public readonly string SpawnsActor = "UnclaimedColony";
+
+		public readonly string NewOwner = "Neutral";
+
+		public readonly CVec Offset = new CVec(0, 0);
 
 		public override object Create(ActorInitializer init)
 		{
-			return new Colony(init, this);
-		}
-
-		public WeaponInfo WeaponInfo { get; private set; }
-		public WeaponInfo ExplodeWeaponInfo { get; private set; }
-
-		public void RulesetLoaded(Ruleset rules, ActorInfo ai)
-		{
-			if (!string.IsNullOrEmpty(Weapon))
-			{
-				var weaponToLower = Weapon.ToLowerInvariant();
-				if (!rules.Weapons.TryGetValue(weaponToLower, out var weapon))
-					throw new YamlException("Weapons Ruleset does not contain an entry '{0}'".F(weaponToLower));
-				WeaponInfo = weapon;
-			}
-
-			if (!string.IsNullOrEmpty(Explode))
-			{
-				var explodesWeaponToLower = Explode.ToLowerInvariant();
-				if (!rules.Weapons.TryGetValue(explodesWeaponToLower, out var explodesInfo))
-					throw new YamlException("Weapons Ruleset does not contain an entry '{0}'".F(explodesWeaponToLower));
-				ExplodeWeaponInfo = explodesInfo;
-			}
+			return new Colony(this, init.Self);
 		}
 	}
 
-	public class Colony : INotifyKilled, ITick
+	public class Colony : INotifyKilled
 	{
 		readonly ColonyInfo info;
-		readonly Health health;
-		readonly Dictionary<Player, int> bitPickers = new Dictionary<Player, int>();
-		readonly RallyPoint rallyPoint;
-		int fireBitTimer;
+		readonly IEnumerable<Turreted> turreted;
 
-		public Colony(ActorInitializer init, ColonyInfo info)
+		public Colony(ColonyInfo info, Actor self)
 		{
 			this.info = info;
-			health = init.Self.Trait<Health>();
-			health.RemoveOnDeath = false; // TODO: Replace this hack.
-			rallyPoint = init.Self.TraitOrDefault<RallyPoint>();
+			turreted = self.TraitsImplementing<Turreted>();
 		}
 
 		void INotifyKilled.Killed(Actor self, AttackInfo e)
 		{
+			if (!self.IsInWorld)
+				return;
+
 			Game.Sound.Play(SoundType.World, info.LostSound, self.CenterPosition);
-			self.ChangeOwner(self.World.Players.First(player => player.InternalName == "Creeps"));
-			self.CancelActivity(); // Stop shooting, we are friends now.
-			CancelProductions(self);
-			fireBitTimer = info.BitFireDelay;
-		}
 
-		List<CPos> GetBitTargetTiles(Actor self)
-		{
-			var cell = self.World.Map.CellContaining(self.CenterPosition);
-			var tiles = new List<CPos>();
-			for (var i = 0; tiles.Count < info.NumberOfBits; i++)
+			var td = new TypeDictionary
 			{
-				tiles = self.World.Map
-					.FindTilesInAnnulus(cell, info.MinRange, info.MaxRange + i)
-					.Where(tilePosition =>
-					{
-						if (!self.World.Map.Contains(tilePosition))
-							return false;
-
-						if (self.World.Map.GetTerrainInfo(tilePosition).Type == "Water")
-							return false;
-
-						if (self.World.WorldActor.Trait<BuildingInfluence>().GetBuildingAt(tilePosition) != null)
-							return false;
-
-						return !self.World.ActorMap.AnyActorsAt(tilePosition);
-					}).ToList();
-			}
-
-			return tiles;
-		}
-
-		private void LaunchBits(Actor self, CPos tile)
-		{
-			var projectile = new ProjectileArgs
-			{
-				Weapon = info.WeaponInfo,
-				Source = self.CenterPosition,
-				SourceActor = self,
-				PassiveTarget = self.World.Map.CenterOfCell(tile),
-				RangeModifiers = new int[0],
-				DamageModifiers = new int[0],
-				InaccuracyModifiers = new int[0]
+				new ParentActorInit(self),
+				new LocationInit(self.Location + info.Offset),
+				new CenterPositionInit(self.CenterPosition),
+				new OwnerInit(info.NewOwner),
 			};
-			self.World.AddFrameEndTask(world => world.Add(projectile.Weapon.Projectile.Create(projectile)));
-		}
 
-		void Explodes(Actor self)
-		{
-			// Don't use Target.FromActor as the actor suicides.
-			info.ExplodeWeaponInfo.Impact(Target.FromPos(self.CenterPosition), self);
-			Game.Sound.Play(SoundType.World, info.ColonyExplosionSound, self.CenterPosition);
+			foreach (var t in turreted)
+				td.Add(new TurretFacingInit(t.Info, t.LocalOrientation.Yaw));
+
+			self.World.AddFrameEndTask(w => w.CreateActor(info.SpawnsActor, td));
 		}
 
 		public void CancelProductions(Actor self)
@@ -143,70 +71,6 @@ namespace OpenRA.Mods.OpenSA.Traits
 						productionQueue.EndProduction(productionItem);
 				}
 			}
-		}
-
-		public void PickBit(Player player)
-		{
-			if (!bitPickers.ContainsKey(player))
-				bitPickers.Add(player, 0);
-
-			bitPickers[player]++;
-		}
-
-		void ITick.Tick(Actor self)
-		{
-			if (fireBitTimer > 0)
-			{
-				if (--fireBitTimer == 0)
-				{
-					var tiles = GetBitTargetTiles(self);
-
-					for (var i = 0; i < info.NumberOfBits; i++)
-					{
-						var tile = tiles[self.World.SharedRandom.Next(0, tiles.Count)];
-						tiles.Remove(tile);
-						LaunchBits(self, tile);
-						Explodes(self);
-					}
-				}
-			}
-
-			if (!self.IsDead)
-				return;
-
-			if (bitPickers.Values.Sum() < info.NumberOfBits)
-				return;
-
-			var bits = bitPickers.Aggregate((a, b) => a.Value > b.Value ? a : b).Value;
-			var newOwner = bitPickers.First(entry => entry.Value == bits).Key;
-
-			if (bitPickers.Values.Count(value => value == bits) > 1)
-				newOwner = self.World.Players.First(player => player.InternalName == "Creeps");
-
-			self.ChangeOwner(newOwner);
-			Game.Sound.Play(SoundType.World, info.CaptureSound, self.CenterPosition);
-			bitPickers.Clear();
-
-			// Get the timing right as owner change is also done at the end of the frame.
-			self.World.AddFrameEndTask(_ =>
-			{
-				self.CancelActivity(); // Stop shooting!
-
-				health.Resurrect(self, self);
-				health.InflictDamage(self, self, new Damage(health.MaxHP - info.ResurrectHealth), true);
-
-				// HACK: Workaround due to killing the colony
-				if (rallyPoint != null)
-				{
-					// Display only the first level of priority
-					var priorityExits = self.Info.TraitInfos<ExitInfo>()
-						.GroupBy(e => e.Priority)
-						.FirstOrDefault();
-
-					var exits = priorityExits != null ? priorityExits.ToArray() : new ExitInfo[0];
-					self.World.Add(new RallyPointIndicator(self, rallyPoint, exits));
-				}
-			});
 		}
 	}
 }
