+++ OpenRA.Mods.OpenSA/Traits/World/PirateSpawner.cs
@@ -108,9 +108,8 @@ namespace OpenRA.Mods.OpenSA.Traits
 					continue;
 
 				// Don't drop on any actors
-				if (self.World.WorldActor.Trait<BuildingInfluence>().GetBuildingAt(p) != null
-					|| self.World.ActorMap.GetActorsAt(p).Any())
-						continue;
+				if (self.World.ActorMap.GetActorsAt(p).Any())
+					continue;
 
 				return p;
 			}
