+++ mod.config
@@ -9,7 +9,7 @@
 MOD_ID="sa"
 
 # The OpenRA engine version to use for this project.
-ENGINE_VERSION="da53d5b77699fbf996d9e86dbd60f9c6af1390df"
+ENGINE_VERSION="72c82cb080500d0b24c5313ff6baf3161b67dc4e"
 
 # .dll filenames compiled by the mod solution for use by the runtime assembly check
 WHITELISTED_MOD_ASSEMBLIES="OpenRA.Mods.OpenSA.dll"
