+++ mods/sa/rules/ants-buildings.yaml
@@ -71,6 +71,58 @@ ants_colony:
 	WithLoopedAttackSound:
 		Report: sounds|TURRETMACHINEGUN.SDF
 		Delay: 6
+	Colony:
+		SpawnsActor: defeated_ants_colony
+
+defeated_ants_colony:
+	Inherits: ^CoreUnclaimedBuilding
+	DefeatedColony:
+		SpawnsActor: ants_colony
+	RenderSprites:
+		Image: ants_colony
+	Turreted@primary:
+		Turret: primary
+		Offset: 170,-140,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	WithSpriteTurret@primary:
+		Turret: primary
+	Armament@primary:
+		Name: primary
+		Turret: primary
+		Weapon: ants_colony
+	WithTurretAttackAnimation@primary:
+		Armament: primary
+		Turret: primary
+		Sequence: shoot
+	# Turret 2
+	Turreted@secondary:
+		Turret: secondary
+		Offset: 3850,500,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	WithSpriteTurret@secondary:
+		Turret: secondary
+	Armament@secondary:
+		Name: secondary
+		Turret: secondary
+		Weapon: ants_colony
+	WithTurretAttackAnimation@secondary:
+		Armament: secondary
+		Turret: secondary
+		Sequence: shoot
+	AttackTurreted:
+		Turrets: primary, secondary
+	Building:
+		Dimensions: 6,6
+		Footprint: =xxx== xxxxx= xxxxx= xxxxx= ==xx== ======
+	HitShape:
+		Type: Rectangle
+			TopLeft: -1536, -2048
+			BottomRight: 1536, 1024
+		TargetableOffsets: -819,-1024,0, -819,0,0, -819,1024,0, 819,-1024,0, 819,0,0, 819,1024,0, 2458,-1024,0, 2458,0,0, 2458,1024,0 # Y,X,Z - Y uses dark space magic. do NOT touch or kittens will die!
 
 ants_colony_turret:
 	Inherits@1: ^CoreTurretBuilding
@@ -107,3 +159,30 @@ ants_colony_turret:
 	WithLoopedAttackSound:
 		Report: sounds|TURRETMACHINEGUN.SDF
 		Delay: 6
+	Colony:
+		SpawnsActor: defeated_ants_turret
+
+defeated_ants_turret:
+	Inherits: ^CoreUnclaimedTurretBuilding
+	DefeatedColony:
+		SpawnsActor: ants_colony_turret
+	RenderSprites:
+		Image: ants_colony_turret
+	Turreted:
+		Turret: primary
+		Offset: 1100,10,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	WithSpriteTurret:
+		Turret: primary
+	Armament:
+		Name: primary
+		Turret: primary
+		Weapon: ants_colony
+	WithTurretAttackAnimation:
+		Armament: primary
+		Turret: primary
+		Sequence: shoot
+	AttackTurreted:
+		Turrets: primary
