+++ mods/sa/rules/beetles-buildings.yaml
@@ -41,6 +41,32 @@ beetles_colony:
 		TurnSpeed: 120
 		InitialFacing: 500
 		RealignDelay: -1
+	Colony:
+		SpawnsActor: defeated_beetles_colony
+
+defeated_beetles_colony:
+	Inherits: ^CoreUnclaimedBuilding
+	Inherits@2: ^CoreTurret
+	DefeatedColony:
+		SpawnsActor: beetles_colony
+	RenderSprites:
+		Image: beetles_colony
+	Armament:
+		Weapon: beetles_colony
+		LocalOffset: 750,0,0
+	Turreted:
+		Offset: 1400,0,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	Building:
+		Dimensions: 6,6
+		Footprint: ==xxx= =xxxx= xxxxxx xxxxxx =xxx== ==xx==
+	HitShape:
+		Type: Rectangle
+			TopLeft: -1536, -2048
+			BottomRight: 1536, 1024
+		TargetableOffsets: -819,-1024,0, -819,0,0, -819,1024,0, 819,-1024,0, 819,0,0, 819,1024,0, 2458,-1024,0, 2458,0,0, 2458,1024,0 # Y,X,Z - Y uses dark space magic. do NOT touch or kittens will die!
 
 beetles_colony_turret:
 	Inherits@1: ^CoreTurretBuilding
@@ -56,6 +82,34 @@ beetles_colony_turret:
 		Type: Circle
 			Radius: 512
 	WithIdleOverlay:
+	Turreted:
+		Turret: primary
+		Offset: 100,100,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	WithSpriteTurret:
+		Turret: primary
+	Armament:
+		Name: primary
+		Turret: primary
+		Weapon: beetles_colony
+		LocalOffset: 750,0,0
+	WithTurretAttackAnimation:
+		Armament: primary
+		Turret: primary
+		Sequence: shoot
+	AttackTurreted:
+		Turrets: primary
+	Colony:
+		SpawnsActor: defeated_beetles_turret
+
+defeated_beetles_turret:
+	Inherits: ^CoreUnclaimedTurretBuilding
+	DefeatedColony:
+		SpawnsActor: beetles_colony_turret
+	RenderSprites:
+		Image: beetles_colony_turret
 	Turreted:
 		Turret: primary
 		Offset: 100,100,0
