+++ mods/sa/rules/core.yaml
@@ -44,20 +44,10 @@
 		TargetTypes: Ground, Colony
 	HitShape:
 	Colony:
-		BitFireDelay: 75 #Time before colony bits are launched
-		NumberOfBits: 9 #How many colony bits are launched
-		MinRange: 5 #Min range of colony bits
-		MaxRange: 10 #Max range of colony bits
-		ResurrectHealth: 50000
-		Weapon: colony_bit
-		CaptureSound: sounds|POWERUP.SDF
-		LostSound: sounds|POWERDOWN.SDF
 	VerticalSelectionDecorations:
 	Selectable:
 		Priority: 10
 	WithSpriteBody:
-	MustBeDestroyed:
-		RequiredForShortGame: true
 	# Group all buildings in the map editor.
 	MapEditorData:
 		Categories: Building
@@ -80,12 +70,38 @@
 		Cursor: move
 		Path: 5,5
 		LineWidth: 3
-	GrantConditionOnHealth:
-		Condition: HoldFire
-		MinHP: 0
-		MaxHP: 1
-	AttackTurreted:
-		RequiresCondition: !HoldFire
+
+^CoreUnclaimedBuilding:
+	Inherits: ^CoreVisibleActor
+	Tooltip:
+		Name: Unclaimed Colony
+	HitShape:
+	Health:
+		HP: 50000
+	# Last seen state of a building will be rendered till the area is explored again.
+	-HiddenUnderFog:
+	FrozenUnderFog:
+	Building:
+		TerrainTypes: Clear
+	Interactable:
+	WithSpriteBody:
+	WithIdleOverlay:
+	DamageMultiplier@INVULNERABLE:
+		Modifier: 0
+	DefeatedColony:
+		BitFireDelay: 75 #Time before colony bits are launched
+		NumberOfBits: 9 #How many colony bits are launched
+		MinRange: 5 #Min range of colony bits
+		MaxRange: 10 #Max range of colony bits
+		ResurrectHealth: 50
+		Weapon: colony_bit
+		CaptureSound: sounds|POWERUP.SDF
+	CustomAppearsOnRadar:
+	# Actor can be targeted by anything which attacks ground.
+	Targetable:
+		TargetTypes: Ground, Colony
+	DamageMultiplier@INVULNERABLE:
+		Modifier: 0
 
 ^CoreTurretBuilding:
 	Inherits: ^CoreVisibleActor
@@ -100,16 +116,6 @@
 		TargetTypes: Ground, Colony
 	HitShape:
 	Colony:
-		BitFireDelay: 50 #Time before colony bits are launched
-		NumberOfBits: 4 #How many colony bits are launched
-		MinRange: 3 #Min range of colony bits
-		MaxRange: 7 #Max range of colony bits
-		ResurrectHealth: 25000
-		Weapon: colony_bit
-		CaptureSound: sa|bits/sounds/TurretPowerUp.wav
-		LostSound: sa|bits/sounds/TurretPowerDown.wav
-		ColonyExplosionSound: sa|bits/sounds/TurretExplosion.wav
-		Explode: dieTurret
 	VerticalSelectionDecorations:
 	Selectable:
 		Priority: 10
@@ -137,12 +143,38 @@
 	WithTurretAttackAnimation:
 		Sequence: shoot
 	AppearsOnRadar:
-	GrantConditionOnHealth:
-		Condition: HoldFire
-		MinHP: 0
-		MaxHP: 1
-	AttackTurreted:
-		RequiresCondition: !HoldFire
+
+^CoreUnclaimedTurretBuilding:
+	Inherits: ^CoreVisibleActor
+	Interactable:
+	Tooltip:
+		Name: Unclaimed Turret
+	# Last seen state of a building will be rendered till the area is explored again.
+	-HiddenUnderFog:
+	FrozenUnderFog:
+	# Use building specific logic.
+	Building:
+		TerrainTypes: Clear
+	HitShape:
+	Health:
+		HP: 50000
+	WithSpriteBody:
+	WithIdleOverlay:
+	DefeatedColony:
+		BitFireDelay: 50 #Time before colony bits are launched
+		NumberOfBits: 4 #How many colony bits are launched
+		MinRange: 3 #Min range of colony bits
+		MaxRange: 7 #Max range of colony bits
+		ResurrectHealth: 50
+		Weapon: colony_bit
+		CaptureSound: sa|bits/sounds/TurretPowerUp.wav
+		ColonyExplosionSound: sa|bits/sounds/TurretExplosion.wav
+		Explode: dieTurret
+	# Actor can be targeted by anything which attacks ground.
+	Targetable:
+		TargetTypes: Ground, Colony
+	DamageMultiplier@INVULNERABLE:
+		Modifier: 0
 
 ^CoreUnit:
 	Inherits: ^CoreCreature
