+++ mods/sa/rules/misc.yaml
@@ -139,7 +139,6 @@ World:
 	PlayerCommands:
 	HelpCommand:
 	ScreenShaker:
-	BuildingInfluence:
 	ProductionQueueFromSelection:
 		ProductionPaletteWidget: PRODUCTION_PALETTE
 	DomainIndex:
