+++ mods/sa/rules/scorpions-buildings.yaml
@@ -46,6 +46,37 @@ scorpions_colony:
 		TurnSpeed: 120
 		InitialFacing: 500
 		RealignDelay: -1
+	Colony:
+		SpawnsActor: defeated_scorpions_colony
+
+defeated_scorpions_colony:
+	Inherits@1: ^CoreUnclaimedBuilding
+	Inherits@2: ^CoreTurret
+	DefeatedColony:
+		SpawnsActor: scorpions_colony
+	RenderSprites:
+		Image: scorpions_colony
+	Armament@primary:
+		Name: primary
+		Weapon: scorpions_colony
+		LocalOffset: 0,256,0
+	Armament@secondary:
+		Name: secondary
+		Weapon: scorpions_colony
+		LocalOffset: 0,-256,0
+	Turreted:
+		Offset: 3450,440,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	Building:
+		Dimensions: 6,6
+		Footprint: =xxxx= xxxxxx xxxxxx xxxxxx =xxxxx ===x==
+	HitShape:
+		Type: Rectangle
+			TopLeft: -1536, -2048
+			BottomRight: 1536, 1024
+		TargetableOffsets: -819,-1024,0, -819,0,0, -819,1024,0, 819,-1024,0, 819,0,0, 819,1024,0, 2458,-1024,0, 2458,0,0, 2458,1024,0 # Y,X,Z - Y uses dark space magic. do NOT touch or kittens will die!
 
 scorpions_colony_turret:
 	Inherits@1: ^CoreTurretBuilding
@@ -61,6 +92,29 @@ scorpions_colony_turret:
 		Type: Circle
 			Radius: 512
 	WithIdleOverlay:
+	Turreted:
+		Offset: 800,-50,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	Armament@primary:
+		Name: primary
+		Weapon: scorpions_colony
+		LocalOffset: 0,256,0
+	Armament@secondary:
+		Name: secondary
+		Weapon: scorpions_colony
+		LocalOffset: 0,-256,0
+	Colony:
+		SpawnsActor: defeated_scorpions_turret
+
+defeated_scorpions_turret:
+	Inherits@1: ^CoreUnclaimedTurretBuilding
+	Inherits@2: ^CoreTurret
+	DefeatedColony:
+		SpawnsActor: scorpions_colony_turret
+	RenderSprites:
+		Image: scorpions_colony_turret
 	Turreted:
 		Offset: 800,-50,0
 		TurnSpeed: 120
