+++ mods/sa/rules/spiders-buildings.yaml
@@ -66,9 +66,61 @@ spiders_colony:
 		Armament: secondary
 		Turret: secondary
 		Sequence: shoot
-	
 	AttackTurreted:
 		Turrets: primary, secondary
+	Colony:
+		SpawnsActor: defeated_spiders_colony
+
+defeated_spiders_colony:
+	Inherits: ^CoreUnclaimedBuilding
+	DefeatedColony:
+		SpawnsActor: spiders_colony
+	RenderSprites:
+		Image: spiders_colony
+	# Turret 1
+	Turreted@primary:
+		Turret: primary
+		Offset: 800,950,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	WithSpriteTurret@primary:
+		Turret: primary
+	Armament@primary:
+		Name: primary
+		Turret: primary
+		Weapon: spiders_colony
+	WithTurretAttackAnimation@primary:
+		Armament: primary
+		Turret: primary
+		Sequence: shoot
+	# Turret 2
+	Turreted@secondary:
+		Turret: secondary
+		Offset: 3200,-850,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	WithSpriteTurret@secondary:
+		Turret: secondary
+	Armament@secondary:
+		Name: secondary
+		Turret: secondary
+		Weapon: spiders_colony
+	WithTurretAttackAnimation@secondary:
+		Armament: secondary
+		Turret: secondary
+		Sequence: shoot
+	AttackTurreted:
+		Turrets: primary, secondary
+	Building:
+		Dimensions: 6,6
+		Footprint: =xxxx= =xxxxx xxxxxx xxxxxx xxxx== ==xx==
+	HitShape:
+		Type: Rectangle
+			TopLeft: -1536, -2048
+			BottomRight: 1536, 1024
+		TargetableOffsets: -819,-1024,0, -819,0,0, -819,1024,0, 819,-1024,0, 819,0,0, 819,1024,0, 2458,-1024,0, 2458,0,0, 2458,1024,0 # Y,X,Z - Y uses dark space magic. do NOT touch or kittens will die!
 
 spiders_colony_turret:
 	Inherits@1: ^CoreTurretBuilding
@@ -84,6 +136,33 @@ spiders_colony_turret:
 		Type: Circle
 			Radius: 256
 	WithIdleOverlay:
+	Turreted:
+		Turret: primary
+		Offset: 1000,0,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	WithSpriteTurret:
+		Turret: primary
+	Armament:
+		Name: primary
+		Turret: primary
+		Weapon: spiders_colony
+	WithTurretAttackAnimation:
+		Armament: primary
+		Turret: primary
+		Sequence: shoot
+	AttackTurreted:
+		Turrets: primary
+	Colony:
+		SpawnsActor: defeated_spiders_turret
+
+defeated_spiders_turret:
+	Inherits: ^CoreUnclaimedTurretBuilding
+	DefeatedColony:
+		SpawnsActor: spiders_colony_turret
+	RenderSprites:
+		Image: spiders_colony_turret
 	Turreted:
 		Turret: primary
 		Offset: 1000,0,0
