+++ mods/sa/rules/wasps-buildings.yaml
@@ -46,6 +46,32 @@ wasps_colony:
 		TurnSpeed: 120
 		InitialFacing: 500
 		RealignDelay: -1
+	Colony:
+		SpawnsActor: defeated_wasps_colony
+
+defeated_wasps_colony:
+	Inherits@1: ^CoreUnclaimedBuilding
+	Inherits@2: ^CoreTurret
+	DefeatedColony:
+		SpawnsActor: wasps_colony
+	RenderSprites:
+		Image: wasps_colony
+	Armament:
+		Weapon: wasps_colony
+		LocalOffset: 750,0,0
+	Turreted:
+		Offset: 2150,-950,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	Building:
+		Dimensions: 6,6
+		Footprint: =xxxx= xxxxxx xxxxxx xxxxxx =xxxxx ======
+	HitShape:
+		Type: Rectangle
+			TopLeft: -1536, -2048
+			BottomRight: 1536, 1024
+		TargetableOffsets: -819,-1024,0, -819,0,0, -819,1024,0, 819,-1024,0, 819,0,0, 819,1024,0, 2458,-1024,0, 2458,0,0, 2458,1024,0 # Y,X,Z - Y uses dark space magic. do NOT touch or kittens will die!
 
 wasps_colony_turret:
 	Inherits@1: ^CoreTurretBuilding
@@ -64,6 +90,34 @@ wasps_colony_turret:
 	WithLoopedAttackSound:
 		Report: sounds|CHAINGUN.SDF
 		Delay: 4
+	Turreted:
+		Turret: primary
+		Offset: 1100,0,0
+		TurnSpeed: 120
+		InitialFacing: 500
+		RealignDelay: -1
+	WithSpriteTurret:
+		Turret: primary
+	Armament:
+		Name: primary
+		Turret: primary
+		Weapon: wasps_colony
+		LocalOffset: 750,0,0
+	WithTurretAttackAnimation:
+		Armament: primary
+		Turret: primary
+		Sequence: shoot
+	AttackTurreted:
+		Turrets: primary
+	Colony:
+		SpawnsActor: defeated_wasps_turret
+
+defeated_wasps_turret:
+	Inherits: ^CoreUnclaimedTurretBuilding
+	DefeatedColony:
+		SpawnsActor: wasps_colony_turret
+	RenderSprites:
+		Image: wasps_colony_turret
 	Turreted:
 		Turret: primary
 		Offset: 1100,0,0
